Kebab/pyproject.toml
[tool.poetry]
name = "kebab"
version = "0.1.0"
description = ""
authors = ["aitorru <aitorruiz2m@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
transformers = "^4.42.3"
torch = "^2.3.1"
tiktoken = "^0.7.0"
einops = "^0.8.0"
pytest = "^8.2.2"
llama-cpp-python = "^0.2.82"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
